+++ shop.py
@@ -43,10 +43,10 @@ shop_items = {
 
 # These upgraeds can be renamed to whatever is necesarry to match the shop items, we can also change the cps and cpc values to whatever we want
 shop_upgrades = {
-    'Click Multiplier 1': ShopUpgrade("Click Multiplier 1", 200, None, 1.05, 'assets/in_game_buttons/click_multipier_1_rectangle.png'),  # Can rename this to rolling pin 1
-    'Click Multiplier 2': ShopUpgrade("Click Multiplier 2", 500, None, 1.15, 'assets/in_game_buttons/click_multiplier_2_rectangle.png'),  # Can rename this to rolling pin 2
-    'Click Multiplier 3': ShopUpgrade("Click Multiplier 3", 1000, None, 1.35, 'assets/in_game_buttons/click_multiplier_3_rectangle.png'),  # Can rename this to rolling pin 3
-    'Increase Click 1': ShopUpgrade("Increase Click 1", 300, None, 2, 'assets/in_game_buttons/increase_click_1_rectangle.png'),  # Can rename this to Reinforced Hands
-    'Increase Click 2': ShopUpgrade("Increase Click 2", 600, None, 3, 'assets/in_game_buttons/increase_click_2_rectangle.png'),  # Can rename this to Strengthened Hands
-    'Increase Click 3': ShopUpgrade("Increase Click 3", 1200, None, 5, 'assets/in_game_buttons/increase_click_3_rectangle.png')  # Can rename this to Sturdy Hands
+    'Click Multiplier 1': ShopUpgrade("Click Multiplier 1", 2, None, 1.05, 'assets/in_game_buttons/click_multipier_1_rectangle.png'),  # Can rename this to rolling pin 1
+    'Click Multiplier 2': ShopUpgrade("Click Multiplier 2", 5, None, 1.15, 'assets/in_game_buttons/click_multiplier_2_rectangle.png'),  # Can rename this to rolling pin 2
+    'Click Multiplier 3': ShopUpgrade("Click Multiplier 3", 10, None, 1.35, 'assets/in_game_buttons/click_multiplier_3_rectangle.png'),  # Can rename this to rolling pin 3
+    'Increase Click 1': ShopUpgrade("Increase Click 1", 3, None, 2, 'assets/in_game_buttons/increase_click_1_rectangle.png'),  # Can rename this to Reinforced Hands
+    'Increase Click 2': ShopUpgrade("Increase Click 2", 6, None, 3, 'assets/in_game_buttons/increase_click_2_rectangle.png'),  # Can rename this to Strengthened Hands
+    'Increase Click 3': ShopUpgrade("Increase Click 3", 12, None, 5, 'assets/in_game_buttons/increase_click_3_rectangle.png')  # Can rename this to Sturdy Hands
 }